Implement BundleService getExternalBundleFolder Change appId type from Long to Integer

+++ ZaloPay/data/src/main/java/vn/com/vng/zalopay/data/cache/SqlitePlatformScope.java
@@ -24,7 +24,7 @@ public interface SqlitePlatformScope extends SqlBaseScope {
     List<AppResourceEntity> listAppResourceEntity();
 
     // Xoá app, nếu không tồn tại trong list này.
-    void updateAppId(List<Long> list);
+    void updateAppId(List<Integer> list);
 
     void writePaymentTransType(List<PaymentTransTypeEntity> list);
